stp/tsp/requirements.txt
Django==2.2.7
djangorestframework==3.11.0
mysqlclient==1.4.6
Pillow==6.2.1
pytz==2019.3
qrcode==6.1
six==1.13.0
sqlparse==0.3.0
